Fixed WaiDaemonsetReady to continue checking if DesiredNumberSchedued is not valid.
In some cases, the daemonset's field DesiredNumberScheduled takes a bit
longer to be updated to the right value by k8s, so it remains as 0 for a
short period of time. This function has been updated to wait instead
of returning an error when the values don't match.

This change is similar to this TNF's one:
https://github.com/test-network-function/cnf-certification-test/pull/253